Split out the utils into separate files
